GameFactory can create all types of games. But logic in AddCommand use instanceof for using builder of game child type

+++ src/GamesCollection/Command/DeleteAllGamesCommand.java
@@ -5,6 +5,11 @@ import GamesCollection.GameUtil;
 public class DeleteAllGamesCommand implements Command{
     @Override
     public void execute() {
-        GameUtil.deleteAllGames();
+        if(GameUtil.deleteAllGames()){
+            GameUtil.deleteAllGames();
+            System.out.println("Все игры удалены из коллекции");
+        }else {
+            System.out.println("Коллекция пуста, удалять пока нечего");
+        }
     }
 }
